@@ -23,75 +23,77 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.1.4",
-    "@angular/cdk": "6.4.6",
-    "@angular/common": "6.1.4",
-    "@angular/compiler": "6.1.4",
-    "@angular/core": "6.1.4",
-    "@angular/flex-layout": "6.0.0-beta.16",
-    "@angular/forms": "6.1.4",
-    "@angular/http": "6.1.4",
-    "@angular/material": "6.4.6",
-    "@angular/material-moment-adapter": "6.4.6",
-    "@angular/platform-browser": "6.1.4",
-    "@angular/platform-browser-dynamic": "6.1.4",
-    "@angular/platform-server": "6.1.4",
-    "@angular/router": "6.1.4",
-    "@ngrx/effects": "6.0.1",
-    "@ngrx/entity": "6.0.1",
-    "@ngrx/router-store": "6.0.1",
-    "@ngrx/store": "6.0.1",
-    "@ngrx/store-devtools": "6.0.1",
-    "@nguniversal/common": "v6.1.0",
-    "@nguniversal/express-engine": "v6.1.0",
-    "@nguniversal/module-map-ngfactory-loader": "v6.1.0",
-    "angular-calendar": "0.25.2",
-    "core-js": "2.5.3",
+    "@angular/animations": "7.1.0",
+    "@angular/cdk": "7.1.0",
+    "@angular/common": "7.1.0",
+    "@angular/compiler": "7.1.0",
+    "@angular/core": "7.1.0",
+    "@angular/flex-layout": "7.0.0-beta.19",
+    "@angular/forms": "7.1.0",
+    "@angular/http": "7.1.0",
+    "@angular/material": "7.1.0",
+    "@angular/material-moment-adapter": "7.1.0",
+    "@angular/platform-browser": "7.1.0",
+    "@angular/platform-browser-dynamic": "7.1.0",
+    "@angular/platform-server": "7.1.0",
+    "@angular/router": "7.1.0",
+    "@ngrx/effects": "7.0.0-beta.0",
+    "@ngrx/entity": "7.0.0-beta.0",
+    "@ngrx/router-store": "7.0.0-beta.0",
+    "@ngrx/store": "7.0.0-beta.0",
+    "@ngrx/store-devtools": "7.0.0-beta.0",
+    "@nguniversal/common": "7.0.2",
+    "@nguniversal/express-engine": "7.0.2",
+    "@nguniversal/module-map-ngfactory-loader": "7.0.2",
+    "angular-calendar": "0.26.4",
+    "core-js": "2.5.7",
     "crypto-js": "3.1.9-1",
     "date-fns": "1.29.0",
-    "express": "4.16.2",
+    "express": "4.16.4",
     "hammerjs": "2.0.8",
-    "lodash": "4.17.5",
-    "moment": "2.20.1",
+    "lodash": "4.17.11",
+    "moment": "2.22.2",
     "rxjs": "^6.2.2",
-    "ts-loader": "^4.5.0",
+    "ts-loader": "^5.3.1",
+    "tslib": "^1.9.0",
     "web-animations-js": "2.3.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular-devkit/core": "^0.7.4",
-    "@angular-devkit/schematics": "^0.7.4",
-    "@angular/cli": "6.1.4",
-    "@angular/compiler-cli": "6.1.4",
-    "@angular/language-service": "6.1.4",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular-devkit/core": "^7.1.0",
+    "@angular-devkit/schematics": "^7.1.0",
+    "@angular/cli": "7.1.0",
+    "@angular/compiler-cli": "7.1.0",
+    "@angular/language-service": "7.1.0",
     "@angularclass/hmr": "2.1.3",
-    "@types/crypto-js": "3.1.38",
-    "@types/express": "4.11.1",
-    "@types/jasmine": "2.8.6",
+    "@types/crypto-js": "3.1.43",
+    "@types/express": "4.16.0",
+    "@types/jasmine": "3.3.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/lodash": "4.14.104",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.2.1",
-    "concurrently": "3.5.1",
+    "@types/lodash": "4.14.118",
+    "@types/node": "~10.12.10",
+    "codelyzer": "~4.5.0",
+    "concurrently": "4.1.0",
     "http-server": "^0.11.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "0.3.1",
+    "jasmine-core": "~3.3.0",
+    "jasmine-marbles": "0.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "ngrx-store-freeze": "0.2.4",
-    "protractor": "~5.3.0",
-    "rxjs-tslint-rules": "~4.7.2",
-    "ts-node": "~5.0.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.9.2",
-    "udk": "^0.3.17",
+    "protractor": "~5.4.1",
+    "rxjs-tslint-rules": "~4.13.0",
+    "ts-node": "~7.0.1",
+    "tslint": "~5.11.0",
+    "typescript": "3.1.6",
+    "udk": "^1.0.0",
+    "webpack": "^4.26.1",
     "webpack-cli": "^3.1.0",
-    "webpack-node-externals": "1.6.0"
+    "webpack-node-externals": "1.7.2"
   },
   "description": "The Task Manager System",
   "main": "index.js",
