@@ -1,6 +1,15 @@
 import { Component, HostBinding } from '@angular/core';
 import { CalendarEvent } from 'angular-calendar';
-import { addDays, addHours, endOfDay, endOfMonth, isSameDay, isSameMonth, startOfDay, subDays } from 'date-fns';
+import {
+  addDays,
+  addHours,
+  endOfDay,
+  endOfMonth,
+  isSameDay,
+  isSameMonth,
+  startOfDay,
+  subDays
+} from 'date-fns';
 import { Observable } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
 import { ActivatedRoute } from '@angular/router';
@@ -13,76 +22,80 @@ import * as fromRoot from '../../reducers';
   selector: 'app-cal-home',
   template: `
     <mat-card fxLayout="column" fxLayoutAlign="start stretch">
-    <div fxLayout="row">
-      <button
-        mat-icon-button
-        mwlCalendarPreviousView
-        [view]="view$ | async"
-        [(viewDate)]="viewDate">
-        <mat-icon class="md-48">chevron_left</mat-icon>
-      </button>
-      <button
-        mat-button
-        mwlCalendarToday
-        [(viewDate)]="viewDate">
-        {{ viewDate | date: 'yyyy-MM-dd' }}
-      </button>
-      <button
-        mat-icon-button
-        mwlCalendarNextView
-        [view]="view$ | async"
-        [(viewDate)]="viewDate">
-        <mat-icon class="md-48">chevron_right</mat-icon>
-      </button>
-    </div>
-    <div *ngIf="(events$ | async) as calEvents">
-      <div [ngSwitch]="view$ | async">
-        <mwl-calendar-month-view
-          *ngSwitchCase="'month'"
-          [viewDate]="viewDate"
-          [locale]="'zh'"
-          [events]="calEvents"
-          [activeDayIsOpen]="activeDayIsOpen"
-          (dayClicked)="dayClicked($event.day)"
-          (eventClicked)="handleEvent('Clicked', $event.event)">
-        </mwl-calendar-month-view>
-        <mwl-calendar-week-view
-          *ngSwitchCase="'week'"
-          [viewDate]="viewDate"
-          [locale]="'zh'"
-          [events]="calEvents"
-          (eventClicked)="handleEvent('Clicked', $event.event)">
-        </mwl-calendar-week-view>
-        <mwl-calendar-day-view
-          *ngSwitchCase="'day'"
-          [viewDate]="viewDate"
-          [locale]="'zh'"
-          [events]="calEvents"
-          (eventClicked)="handleEvent('Clicked', $event.event)">
-        </mwl-calendar-day-view>
+      <div fxLayout="row">
+        <button
+          mat-icon-button
+          mwlCalendarPreviousView
+          [view]="view$ | async"
+          [(viewDate)]="viewDate"
+        >
+          <mat-icon class="md-48">chevron_left</mat-icon>
+        </button>
+        <button mat-button mwlCalendarToday [(viewDate)]="viewDate">
+          {{ viewDate | date: 'yyyy-MM-dd' }}
+        </button>
+        <button
+          mat-icon-button
+          mwlCalendarNextView
+          [view]="view$ | async"
+          [(viewDate)]="viewDate"
+        >
+          <mat-icon class="md-48">chevron_right</mat-icon>
+        </button>
       </div>
-     </div>
-     </mat-card>
+      <div *ngIf="(events$ | async) as calEvents">
+        <div [ngSwitch]="view$ | async">
+          <mwl-calendar-month-view
+            *ngSwitchCase="'month'"
+            [viewDate]="viewDate"
+            [locale]="'zh-Hans'"
+            [events]="calEvents"
+            [activeDayIsOpen]="activeDayIsOpen"
+            (dayClicked)="dayClicked($event.day)"
+            (eventClicked)="handleEvent('Clicked', $event.event)"
+          >
+          </mwl-calendar-month-view>
+          <mwl-calendar-week-view
+            *ngSwitchCase="'week'"
+            [viewDate]="viewDate"
+            [locale]="'zh-Hans'"
+            [events]="calEvents"
+            (eventClicked)="handleEvent('Clicked', $event.event)"
+          >
+          </mwl-calendar-week-view>
+          <mwl-calendar-day-view
+            *ngSwitchCase="'day'"
+            [viewDate]="viewDate"
+            [locale]="'zh-Hans'"
+            [events]="calEvents"
+            (eventClicked)="handleEvent('Clicked', $event.event)"
+          >
+          </mwl-calendar-day-view>
+        </div>
+      </div>
+    </mat-card>
   `,
-  styles: [`
-    mat-card {
-      width: 100%;
-    }
-  `
+  styles: [
+    `
+      mat-card {
+        width: 100%;
+      }
+    `
   ],
-  animations: [defaultRouteAnim],
+  animations: [defaultRouteAnim]
 })
 export class CalendarHomeComponent {
-
   @HostBinding('@routeAnim') state = 'in';
   viewDate: Date;
   view$: Observable<string>;
   activeDayIsOpen = true;
   events$: Observable<CalendarEvent[]>;
 
-  constructor(private route: ActivatedRoute,
+  constructor(
+    private route: ActivatedRoute,
     private service$: MyCalService,
-    private store$: Store<fromRoot.State>) {
+    private store$: Store<fromRoot.State>
+  ) {
     this.viewDate = new Date();
     this.view$ = this.route.paramMap.pipe(map(p => <string>p.get('view')));
     this.events$ = this.store$.pipe(
@@ -95,9 +108,12 @@ export class CalendarHomeComponent {
     console.log('events handled');
   }
 
-  dayClicked({ date, events }: { date: Date, events: CalendarEvent[] }): void {
+  dayClicked({ date, events }: { date: Date; events: CalendarEvent[] }): void {
     if (isSameMonth(date, this.viewDate)) {
-      if ((isSameDay(this.viewDate, date) && this.activeDayIsOpen === true) || events.length === 0) {
+      if (
+        (isSameDay(this.viewDate, date) && this.activeDayIsOpen === true) ||
+        events.length === 0
+      ) {
         this.activeDayIsOpen = false;
       } else {
         this.activeDayIsOpen = true;
