@@ -36,7 +36,7 @@ import 'core-js/es6/set';
 import 'web-animations-js'; // Run `npm install --save web-animations-js`.
 /** Evergreen browsers require these. **/
 import 'core-js/es6/reflect';
-import 'core-js/es7/reflect';
+
 /** ALL Firefox browsers require the following to support `@angular/animation`. **/
 // import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 /***************************************************************************************************
